:truck: Moved useFiles hook to its own file
